+++ index.js
@@ -1,6 +1,9 @@
 var sockets = require('./lib/sockets');
 
 module.exports = sockets;
-module.exports.createContext = function(url, connOpts) {
+module.exports.createContext = createContext;
+module.exports.createConnection = createContext;
+
+function createContext(url, connOpts) {
   return new sockets.Context(url, connOpts);
 }