+++ CHANGELOG.md
@@ -4,6 +4,7 @@ rabbit.js release notes
 FUTURE
 ------
 
+* [ADDED] - `createConnection` alias of `createContext` method of the exported module
 * [ADDED] - `inspectIntegrity` method to the `Context` connection prototype
 
 v0.5.0